requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.0.200
sentence-transformers>=2.2.2
chromadb>=0.3.21
beautifulsoup4>=4.12.2
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0
geopy>=2.4.1
chardet>=5.0.0
openai>=1.0.0
